+++ wtypes/_spec.py
@@ -1,0 +1,17 @@
+specification = pluggy.HookspecMarker("wtypes")
+implementation = pluggy.HookimplMarker("wtypes")
+manager = pluggy.PluginManager("wtypes")
+
+
+class spec:
+    @specification(firstresult=True)
+    def validate_type(type):
+        "A hook to validate types."
+
+    @specification(firstresult=True)
+    def validate_object(object, schema):
+        "A hook to validate types."
+
+
+manager.add_hookspecs(spec)
+del pluggy
